v0.7: introduce namespace mRICH for all constant

+++ include/Utility.h
@@ -14,11 +14,11 @@ class Utility
 
     std::string get_IdentifiedParticle(int pid);
     std::pair<std::string,std::string> get_misIdentifiedParticle(int pid);
-    int get_indexSpaceX(float vx);
-    int get_indexSpaceY(float vy);
-    int get_indexMomentumP(float px, float py, float pz);
-    // int get_indexMomentumTheta(float px, float py, float pz);
-    // int get_indexMomentumPhi(float px, float py, float pz);
+    int get_indexSpaceX(double vx);
+    int get_indexSpaceY(double vy);
+    int get_indexMomentumP(double px, double py, double pz);
+    // int get_indexMomentumTheta(double px, double py, double pz);
+    // int get_indexMomentumPhi(double px, double py, double pz);
 
   private:
     strMap mParticleIdMap;
